requirements.txt
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.40.0
anyio==4.12.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
cryptography==41.0.7
ctransformers==0.2.27
distro==1.9.0
exceptiongroup==1.3.1
faiss-cpu==1.9.0
filelock==3.19.1
flask==3.1.3
fsspec==2025.10.0
gitdb==4.0.12
gitpython==3.1.46
google-auth==2.48.0
google-genai==1.47.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.2
idna==3.11
importlib-metadata==8.7.1
iniconfig==2.1.0
itsdangerous==2.2.0
jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
loguru==0.7.2
lxml==6.0.2
markdown-it-py==3.0.0
markupsafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
narwhals==2.16.0
networkx==3.2.1
numpy==2.0.2
openai==1.56.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pluggy==1.6.0
protobuf==6.33.5
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyasn1==0.6.2
pyasn1-modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pydeck==0.9.1
pygments==2.19.2
pypdf==5.1.0
pytest==8.3.4
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2025.2
pyyaml==6.0.3
rank-bm25==0.2.2
referencing==0.36.2
regex==2026.1.15
requests==2.32.5
rich==14.3.3
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==5.1.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.50.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.8.0
tokenizers==0.22.2
toml==0.10.2
tomli==2.4.0
torch==2.8.0
tornado==6.5.4
tqdm==4.67.3
transformers==4.57.6
typer==0.23.2
typer-slim==0.23.2
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
websockets==15.0.1
werkzeug==3.1.6
win32-setctime==1.2.0
zipp==3.23.0
